requirement.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Google Gemini AI
google-generativeai==0.3.1

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27

# Environment Variables
python-dotenv==1.0.0

# CORS Support
python-multipart==0.0.6

# Database (SQLite - built-in with Python)
# No additional package needed for SQLite3

# Optional: For better development experience
# python-jose[cryptography]==3.3.0  # For JWT tokens if needed later
# passlib[bcrypt]==1.7.4  # For password hashing if needed later